requirements.txt
fastapi
uvicorn
pydantic
transformers
sentence-transformers
faiss-cpu
numpy
pandas
scikit-learn
torch
python-dotenv
